requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.4
django-crispy-forms==2.3
django-storages==1.14.4
email-reply-parser==0.5.12
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2024.2
